refactor(project): rename UseConfigStore to ConfigStore
